[moveit] Isolate sample script from test script as a standalone program.
